+++ backend/tests/test_vector.c
@@ -290,7 +290,7 @@ static void test_buffer_vector(int index, int generate, int debug) {
         } else {
             text = data[i].data;
         }
-        length = strlen(text);
+        length = (int) strlen(text);
 
         ret = ZBarcode_Encode(symbol, (unsigned char *) text, length);
         assert_zero(ret, "i:%d ZBarcode_Encode(%d) ret %d != 0 %s\n", i, data[i].symbology, ret, symbol->errtxt);
@@ -397,7 +397,7 @@ static void test_upcean_hrt(int index, int debug) {
         }
         symbol->debug |= debug;
 
-        length = strlen(data[i].data);
+        length = (int) strlen(data[i].data);
 
         ret = ZBarcode_Encode(symbol, (unsigned char *) data[i].data, length);
         assert_zero(ret, "i:%d ZBarcode_Encode(%d) ret %d != 0 %s\n", i, data[i].symbology, ret, symbol->errtxt);
@@ -554,7 +554,7 @@ static void test_stacking(int index, int debug) {
         ret = ZBarcode_Encode(symbol, (unsigned char *) data[i].data, length);
         assert_zero(ret, "i:%d ret %d != zero\n", i, ret);
 
-        length2 = strlen(data[i].data2);
+        length2 = (int) strlen(data[i].data2);
         ret = ZBarcode_Encode(symbol, (unsigned char *) data[i].data2, length2);
         assert_zero(ret, "i:%d ret %d != zero\n", i, ret);
 
@@ -754,7 +754,7 @@ static void test_noncomposite_string_x(int index, int debug) {
         symbol->input_mode = UNICODE_MODE;
         symbol->debug |= debug;
 
-        length = strlen(data[i].data);
+        length = (int) strlen(data[i].data);
 
         ret = ZBarcode_Encode(symbol, (unsigned char *) data[i].data, length);
         assert_zero(ret, "i:%d ZBarcode_Encode(%d) ret %d != 0 %s\n", i, data[i].symbology, ret, symbol->errtxt);
@@ -815,7 +815,7 @@ static void test_upcean_whitespace_width(int index, int debug) {
         symbol->whitespace_width = data[i].whitespace_width;
         symbol->debug |= debug;
 
-        length = strlen(data[i].data);
+        length = (int) strlen(data[i].data);
 
         ret = ZBarcode_Encode(symbol, (unsigned char *) data[i].data, length);
         assert_zero(ret, "i:%d ZBarcode_Encode(%d) ret %d != 0 %s\n", i, data[i].symbology, ret, symbol->errtxt);
@@ -1298,7 +1298,7 @@ static void test_height(int index, int generate, int debug) {
         } else {
             text = data[i].data;
         }
-        length = strlen(text);
+        length = (int) strlen(text);
 
         ret = ZBarcode_Encode(symbol, (unsigned char *) text, length);
         assert_equal(ret, data[i].ret, "i:%d ZBarcode_Encode(%s) ret %d != %d (%s)\n", i, testUtilBarcodeName(data[i].symbology), ret, data[i].ret, symbol->errtxt);
